refactor: extract AppState tracker-integration extension (#583, #430)
Eleventh slice of #430 (AppState decomposition). Per #587 design sketch
and codex spec-review's intent-first recommendation: extracts 7
GitHub/Jira tracker-integration delegators — 6 in the main cluster
plus jiraIssueTypes(for:) as a non-adjacent Cluster A — into a new
Sources/BugNarrator/AppState+TrackerIntegration.swift as
`extension AppState { ... }`.

**Cluster A** (AppState.swift lines 147-149, 3 lines, 147 bytes,
SHA-256 f587cc9a4b88013daafb1332409c0f911e63f7f5b245b754901eeac47ace59dc):
- jiraIssueTypes(for:) -> [JiraIssueTypeOption]

**Cluster B** (AppState.swift lines 917-939, 23 lines, 751 bytes,
SHA-256 35eaf7a7f769fe64c5478a03674838c46974f288ecf2df710759ce34cd20fe7e):
- validateGitHubConfiguration() async
- loadGitHubRepositories() async
- validateJiraConfiguration() async
- selectJiraProject(projectID:)
- refreshJiraIssueTypesForSelectedProject() async
- loadJiraIssueTypes(forProjectID:) async

All 7 methods pass the delegator-vs-orchestrator test — pure passthroughs
to `trackerIntegration`, no `self`-scoped await, no @Published mutation,
no branching. 5 of 7 are async but the shape (single `await
collaborator.method()`) is still a delegator.

Non-adjacent extraction protocol from #556 applied: extract both
clusters into new file in original file order (Cluster A first), then
delete in reverse line-number order (917-940 first, then 147-150) so
earlier ranges stay valid.

Full triple-proof stack per cluster:
- Endpoint verify (147 = jiraIssueTypes header, 149 = `}`; 917 =
  validateGitHubConfiguration header, 939 = `}` of loadJiraIssueTypes)
- SHA gate matched at spec-review (against origin) and post-extract
- Per-cluster byte-diff against origin/main: both silent
- Post-deletion double-blank scan: none

**refreshExportHistory()** (now at line 917 post-move) STAYS in class
body — different collaborator (exportHistoryController), solo delegator
without peers, per #587 design sketch rule 3.

Extension inherits @MainActor from AppState per SE-0316 — NOT
re-annotated. New file imports `import Foundation` (matches house style
per #571+; codex noted it's technically dead in this file since the
bodies only reference stdlib + same-module types, but consistent with
prior slices).

No dead-import cleanup — AppState.swift still needs Combine + Foundation.

667 unit tests pass unchanged; docs-sync validation passes.

Pre-build spec reviewed by codex in a fresh session (2 findings folded:
xcodegen target update, include non-adjacent jiraIssueTypes(for:) per
intent-first rule from #590 review 26 — same domain, ~800 lines away).
