+++ Sources/BugNarrator/AppState.swift
@@ -144,10 +144,6 @@ final class AppState: ObservableObject {
         trackerIntegration.jiraIssueTypes
     }
 
-    func jiraIssueTypes(for target: JiraIssueExportTarget) -> [JiraIssueTypeOption] {
-        trackerIntegration.jiraIssueTypes(for: target)
-    }
-
     var isLoadingJiraIssueTypes: Bool {
         trackerIntegration.isLoadingJiraIssueTypes
     }
@@ -912,30 +908,6 @@ final class AppState: ObservableObject {
 
     func removeAPIKey() {
         aiProviderSettings.removeCredential()
-    }
-
-    func validateGitHubConfiguration() async {
-        await trackerIntegration.validateGitHubConfiguration()
-    }
-
-    func loadGitHubRepositories() async {
-        await trackerIntegration.loadGitHubRepositories()
-    }
-
-    func validateJiraConfiguration() async {
-        await trackerIntegration.validateJiraConfiguration()
-    }
-
-    func selectJiraProject(projectID: String) {
-        trackerIntegration.selectJiraProject(projectID: projectID)
-    }
-
-    func refreshJiraIssueTypesForSelectedProject() async {
-        await trackerIntegration.refreshJiraIssueTypesForSelectedProject()
-    }
-
-    func loadJiraIssueTypes(forProjectID projectID: String) async {
-        await trackerIntegration.loadJiraIssueTypes(forProjectID: projectID)
     }
 
     func refreshExportHistory() async {
